+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Puzzle preview and navigation buttons added above freeplay puzzle select list.
+
 ### Fixed
 
 - Initial freeplay puzzle had incorrect index if puzzle titles were not unique.